Added command to get help in ADB command mode
A common complaint, especially from myself, is that the shell gives you
help when you try to run adb shell commands, but purr doesn't. I've
linked up Ctrl-h to try it's best to give help menu for the command in
the query.

It's probably going to be a bit fidgety, but hey. It does the trick most
of the time.