Refactor preview code to be more robust
Two major issues were fixed: the preview should now be able to pad
itself if you are too close to the top of the input, and it should now
detect when it's seeing multiple exact-duplicate lines and report this.

It should now also explicitately note when it can't find a line. This
should never happen, but if it does, it should help with debugging.