Add commands for tagging and untagging
A common friction point is adding and removing tags from the logs.
Removing is especially important to minimize noise, and tags are very
useful for single-app developers. This should make it easier to use purr
for them.

I'm somewhat iffy on how I implemented the code to remove existing tags
from the query, but I'll see if it causes actual breakages.