Add unique mode.
This mode filters out any input that it's already seen. Right now, it
sucks. It's accurate, but it's really slow. I need to get it to
~1000-2000 QPS before I'd consider it fine, and it's closer to ~50.

How to make this work better is not 100% clear to me, lots of paths
forward but all of them suck just a bit...

For now, this isn't in the instructions because I don't want anybody to
use this as a stable feature.