Fix naming issue with bug report functionality
I forgot to put the serial in the naming functionality, so the serial
would often return erroneous information. While I was there, I also
added some code that will return an error in the /tmp folder if it can't
compile the bug report. This is useful for devices where bug report
functionality may be broken.